file upload and download added

+++ controllers/adminController.js
@@ -1,7 +1,9 @@
 const Product = require('./../models/product');
+const fileHelper = require('../utils/file');
 const {validationResult} = require('express-validator');
 
-let getAddProduct = (req, res) => {
+
+let getAddProduct = (req, res, next) => {
     res.render('admin/edit-product', { 
         pageTitle: 'Add Product', 
         path: '/admin/add-product',
@@ -12,18 +14,36 @@ let getAddProduct = (req, res) => {
     });
 }
 
-let postAddProduct = (req, res) => {
+let postAddProduct = (req, res, next) => {
     const title = req.body.title;
-    const imageUrl = req.body.imageUrl;
+    const image = req.file;
     const price = req.body.price;
     const description = req.body.description;
     const errors = validationResult(req);
 
-    if(!errors.isEmpty()) {
+
+    if (!image) {
+        return res.status(422).render('admin/edit-product', {
+            pageTitle: 'Add Product',
+            path: '/admin/add-product',
+            editing: false,
+            hasError: true,
+            product: {
+                title: title,
+                price: price,
+                description: description
+            },
+            errorMessage: 'Attached file is not an image.',
+            validationErrors: []
+        });
+    }
+    
+
+    if (!errors.isEmpty()) {
         console.log(errors.array());
-        return res.status(422).render('admin/edit-product', {  
-            pageTitle: 'Add Products', 
-            path: '/admin/add-product', 
+        return res.status(422).render('admin/edit-product', {
+            pageTitle: 'Add Product',
+            path: '/admin/add-product',
             editing: false,
             hasError: true,
             product: {
@@ -37,36 +57,47 @@ let postAddProduct = (req, res) => {
         });
     }
 
+    const imageUrl = image.path;
 
     const product = new Product({
+        // _id: new mongoose.Types.ObjectId('5badf72403fd8b5be0366e81'),
         title: title,
-        imageUrl: imageUrl,
         price: price,
         description: description,
+        imageUrl: imageUrl,
         userId: req.user
     });
-
     product.save()
-    .then((result) => {
-        console.log('Product created!');
+    .then(result => {
         // console.log(result);
+        console.log('Product created!');
         res.redirect('/admin/products');
     })
-    .catch(err => console.log(err));
+    .catch(err => {
+        // res.redirect('/500');
+        const error = new Error(err);
+        error.httpStatusCode = 500;
+        return next(error);
+    });
 }
 
 
-let getProducts = (req, res) => {
+let getProducts = (req, res, next) => {
     Product.find({userId: req.user._id})
     // .populate('userId', 'name')
     .then((products) => {
         res.render('admin/products', { prods: products, pageTitle: 'Admin Products', path: '/admin/products' });
     })
-    .catch(err => console.log(err));
+    .catch(err => {
+        console.log(err);
+        const error = new Error(err);
+        error.httpStatusCode = 500;
+        return next(error);
+    });
 }
 
 
-let getEditProduct = (req, res) => {
+let getEditProduct = (req, res, next) => {
     const editMode = req.query.edit;
     const prodId = req.params.productId;
 
@@ -88,27 +119,31 @@ let getEditProduct = (req, res) => {
             validationErrors: []
         });
     })
-    .catch(err => console.log(err));
+    .catch(err => {
+        console.log(err);
+        const error = new Error(err);
+        error.httpStatusCode = 500;
+        return next(error);
+    });
 }
 
-let postEditProduct = (req, res) => {
+let postEditProduct = (req, res, next) => {
     const prodId = req.body.productId;
     const title = req.body.title;
-    const imageUrl = req.body.imageUrl;
     const price = req.body.price;
+    const image = req.file;
     const description = req.body.description;
+
     const errors = validationResult(req);
 
     if (!errors.isEmpty()) {
-        // console.log(errors.array());
         return res.status(422).render('admin/edit-product', {
             pageTitle: 'Edit Product',
-            path: '/admin/add-product',
+            path: '/admin/edit-product',
             editing: true,
             hasError: true,
             product: {
                 title: title,
-                imageUrl: imageUrl,
                 price: price,
                 description: description,
                 _id: prodId
@@ -118,35 +153,53 @@ let postEditProduct = (req, res) => {
         });
     }
 
-
     Product.findById(prodId)
     .then(product => {
-        if(product.userId.toString() !== req.user._id.toString()) {
+        if (product.userId.toString() !== req.user._id.toString()) {
             return res.redirect('/');
         }
         product.title = title;
-        product.imageUrl = imageUrl;
         product.price = price;
         product.description = description;
-
-        return product.save()
-        .then((result) => {
-            console.log('Product Updated!');
+        if (image) {
+            fileHelper.deleteFile(product.imageUrl);
+            product.imageUrl = image.path;
+        }
+        return product.save().then(result => {
+            console.log('Product updated!');
             res.redirect('/admin/products');
         });
     })
-    .catch(err => console.log(err));
+    .catch(err => {
+        const error = new Error(err);
+        error.httpStatusCode = 500;
+        return next(error);
+    });
 }
 
 
-let postDeleteProduct = (req, res) => {
+let postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    Product.deleteOne({ _id: prodId, userId: req.user._id })
+
+    Product.findById(prodId)
+    .then(product => {
+        if(!product) {
+            return next(new Error('Product not found.'));
+        }
+        fileHelper.deleteFile(product.imageUrl);
+        
+        return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
     .then(() => {
         console.log('Product deleted');
         res.redirect('/admin/products');
     })
-    .catch(err => console.log(err));
+    .catch(err => {
+        console.log(err);
+        const error = new Error(err);
+        error.httpStatusCode = 500;
+        return next(error);
+    });
 }
